Refactor map and table routes
